requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.0.9
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.1
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-js-asset==1.2.2
djangorestframework==3.13.1
frozenlist==1.2.0
idna==3.3
kombu==5.2.3
multidict==5.2.0
Pillow==9.0.0
prompt-toolkit==3.0.24
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
yarl==1.7.2
